Receive page improvements - payment request button
* Move the "old" tab on the receive page to a button. Remove extra
    frames, tab widget, reorganize and clean up receive page.

* Remove some receive page payment request unit tests

* Add a new popup dialog for payment requests which retains 100% of the
    payment request functionality
