+++ src/qt/receivecoinsdialog.h
@@ -7,14 +7,7 @@
 
 #include <qt/guiutil.h>
 
-#include <QDialog>
-#include <QHeaderView>
-#include <QItemSelection>
-#include <QKeyEvent>
-#include <QMenu>
-#include <QPoint>
-#include <QVariant>
-
+class PaymentRequestDialog;
 class PlatformStyle;
 class WalletModel;
 
@@ -22,64 +15,31 @@ namespace Ui {
     class ReceiveCoinsDialog;
 }
 
-QT_BEGIN_NAMESPACE
-class QModelIndex;
-QT_END_NAMESPACE
-
 /** Dialog for requesting payment of bitcoins */
 class ReceiveCoinsDialog : public QDialog
 {
     Q_OBJECT
 
 public:
-    enum ColumnWidths {
-        DATE_COLUMN_WIDTH = 130,
-        LABEL_COLUMN_WIDTH = 120,
-        AMOUNT_MINIMUM_COLUMN_WIDTH = 180,
-        MINIMUM_COLUMN_WIDTH = 130
-    };
-
     explicit ReceiveCoinsDialog(const PlatformStyle *platformStyle, QWidget *parent = 0);
     ~ReceiveCoinsDialog();
 
     void setModel(WalletModel *model);
 
-public Q_SLOTS:
-    void clear();
-    void reject();
-    void accept();
-
-protected:
-    virtual void keyPressEvent(QKeyEvent *event);
-
 private:
     Ui::ReceiveCoinsDialog *ui;
-    GUIUtil::TableViewLastColumnResizingFixer *columnResizingFixer;
     WalletModel *model;
-    QMenu *contextMenu;
     const PlatformStyle *platformStyle;
 
-    QModelIndex selectedRow();
-    void copyColumnToClipboard(int column);
-    virtual void resizeEvent(QResizeEvent *event);
+    PaymentRequestDialog* requestDialog = nullptr;
 
     void generateQR(std::string data);
     void generateAddress();
 
 private Q_SLOTS:
-    void on_receiveButton_clicked();
-    void on_showRequestButton_clicked();
-    void on_removeRequestButton_clicked();
-    void on_recentRequestsView_doubleClicked(const QModelIndex &index);
-    void recentRequestsView_selectionChanged(const QItemSelection &selected, const QItemSelection &deselected);
-    void updateDisplayUnit();
-    void showMenu(const QPoint &point);
-    void copyURI();
-    void copyLabel();
-    void copyMessage();
-    void copyAmount();
     void on_pushButtonCopy_clicked();
     void on_pushButtonNew_clicked();
+    void on_pushButtonPaymentRequest_clicked();
 };
 
 #endif // BITCOIN_QT_RECEIVECOINSDIALOG_H
